Hardened visibility of methods and fields
