+++ src/cache/manager.rs
@@ -1,26 +1,6 @@
-use crate::cache::redis::{RedisPool, ServiceCache};
-use crate::cache::{Cache, MockCache};
-use crate::config::{
-    default_redis_pool_size, default_redis_uri, info_redis_pool_size, info_redis_uri,
-};
-use r2d2::Pool;
+use crate::cache::redis::ServiceCache;
+use crate::cache::Cache;
 use std::sync::Arc;
-
-fn create_info_pool() -> RedisPool {
-    let client = redis::Client::open(info_redis_uri()).unwrap();
-    Pool::builder()
-        .max_size(info_redis_pool_size())
-        .build(client)
-        .expect("Info redis connection pool initialisation failure")
-}
-
-fn create_default_pool() -> RedisPool {
-    let client = redis::Client::open(default_redis_uri()).unwrap();
-    Pool::builder()
-        .max_size(default_redis_pool_size())
-        .build(client)
-        .expect("Default redis connection pool initialisation failure")
-}
 
 pub trait CacheManager: Sync + Send {
     fn info_cache(&self) -> Arc<dyn Cache>;
@@ -35,13 +15,16 @@ pub struct RedisCacheManager {
 impl RedisCacheManager {
     pub fn new() -> Self {
         RedisCacheManager {
-            info_cache: Arc::new(ServiceCache::new(create_info_pool())),
-            default_cache: Arc::new(ServiceCache::new(create_default_pool())),
+            info_cache: Arc::new(ServiceCache::new_info_cache()),
+            default_cache: Arc::new(ServiceCache::new_default_cache()),
         }
     }
 
     #[cfg(test)]
-    pub fn new_with_mocks(info_cache: MockCache, default_cache: MockCache) -> Self {
+    pub fn new_with_mocks(
+        info_cache: crate::cache::MockCache,
+        default_cache: crate::cache::MockCache,
+    ) -> Self {
         RedisCacheManager {
             info_cache: Arc::new(info_cache) as Arc<dyn Cache>,
             default_cache: Arc::new(default_cache) as Arc<dyn Cache>,
